Fix authentication issue (#5)
* Support V3IO  authentication token and parser in PrestoSQL Hive module

* Implement V3IO dynamic configuration provider

* addressing PR comments
