+++ presto-hive/src/main/java/io/prestosql/plugin/hive/InternalHiveConnectorFactory.java
@@ -29,6 +29,7 @@ import io.prestosql.plugin.hive.metastore.HiveMetastoreModule;
 import io.prestosql.plugin.hive.s3.HiveS3Module;
 import io.prestosql.plugin.hive.security.HiveSecurityModule;
 import io.prestosql.plugin.hive.security.SystemTableAwareAccessControl;
+import io.prestosql.plugin.hive.v3io.HiveV3IOModule;
 import io.prestosql.spi.NodeManager;
 import io.prestosql.spi.PageIndexerFactory;
 import io.prestosql.spi.PageSorter;
@@ -73,6 +74,7 @@ public final class InternalHiveConnectorFactory
                     new HiveModule(),
                     new HiveS3Module(),
                     new HiveGcsModule(),
+                    new HiveV3IOModule(),
                     new HiveMetastoreModule(metastore),
                     new HiveSecurityModule(),
                     new HiveAuthenticationModule(),
