+++ presto-hive/src/main/java/io/prestosql/plugin/hive/v3io/HiveV3IOModule.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.prestosql.plugin.hive.v3io;
+
+import com.google.inject.Binder;
+import com.google.inject.Scopes;
+import io.airlift.configuration.AbstractConfigurationAwareModule;
+import io.prestosql.plugin.hive.DynamicConfigurationProvider;
+
+import static com.google.inject.multibindings.Multibinder.newSetBinder;
+
+public class HiveV3IOModule
+        extends AbstractConfigurationAwareModule
+{
+    @Override
+    protected void setup(Binder binder)
+    {
+        newSetBinder(binder, DynamicConfigurationProvider.class).addBinding().to(V3IOConfigurationProvider.class).in(Scopes.SINGLETON);
+    }
+}
